Separate out GFXQuadList tests and casting tests into their own separate .cpp files
